+++ src/SignalR/clients/csharp/Http.Connections.Client/src/Microsoft.AspNetCore.Http.Connections.Client.csproj
@@ -11,7 +11,8 @@
     <Compile Include="$(SignalRSharedSourceRoot)WebSocketExtensions.cs" Link="WebSocketExtensions.cs" />
     <Compile Include="$(SignalRSharedSourceRoot)StreamExtensions.cs" Link="StreamExtensions.cs" />
     <Compile Include="$(SignalRSharedSourceRoot)DuplexPipe.cs" Link="DuplexPipe.cs" />
-    <Compile Include="$(SignalRSharedSourceRoot)AcknowledgePipeV2.cs" Link="Internal\AcknowledgePipeV2.cs" />
+    <Compile Include="$(SignalRSharedSourceRoot)AckPipeReader.cs" Link="Internal\AckPipeReader.cs" />
+    <Compile Include="$(SignalRSharedSourceRoot)AckPipeWriter.cs" Link="Internal\AckPipeWriter.cs" />
     <Compile Include="$(SignalRSharedSourceRoot)ParseAckPipeReader.cs" Link="Internal\ParseAckPipeReader.cs" />
     <Compile Include="$(SharedSourceRoot)PlatformAttributes.cs" Condition="'$(TargetFramework)' != '$(DefaultNetCoreTargetFramework)'" />
     <Compile Include="$(SharedSourceRoot)OperatingSystem.cs" Condition="'$(TargetFramework)' != '$(DefaultNetCoreTargetFramework)'" />
