+++ src/SignalR/common/SignalR.Common/test/Microsoft.AspNetCore.SignalR.Common.Tests.csproj
@@ -7,7 +7,8 @@
   <ItemGroup>
     <Compile Include="$(SignalRSharedSourceRoot)BinaryMessageFormatter.cs" Link="BinaryMessageFormatter.cs" />
     <Compile Include="$(SignalRSharedSourceRoot)BinaryMessageParser.cs" Link="BinaryMessageParser.cs" />
-    <Compile Include="$(SignalRSharedSourceRoot)AcknowledgePipeV2.cs" Link="Internal\AcknowledgePipeV2.cs" />
+    <Compile Include="$(SignalRSharedSourceRoot)AckPipeReader.cs" Link="Internal\AckPipeReader.cs" />
+    <Compile Include="$(SignalRSharedSourceRoot)AckPipeWriter.cs" Link="Internal\AckPipeWriter.cs" />
     <Compile Include="$(SignalRSharedSourceRoot)ParseAckPipeReader.cs" Link="Internal\ParseAckPipeReader.cs" />
     <Compile Include="$(SharedSourceRoot)Buffers.Testing\**\*.cs" />
   </ItemGroup>
